[FD-1789] Allow Term api_preferences to fallback to Table prefs (#96)
* [FD-1789] Allow Term api_preferences to fallback to Table prefs

* [FD-1789] Update fallback to terminology preferences

* [FD-1789] Update README

* [FD-1789] Update README

* [FD-1789] Update README
